# Back/TEG_Juego/docs/api_doc/swagger.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:55777","description":"Generated server url"}],"paths":{"/api/Usuario/{id}":{"get":{"tags":["usuario-controller"],"operationId":"obtenerPorId","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/usuarioDTO"}}}}}},"put":{"tags":["usuario-controller"],"operationId":"actualizar","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Usuario"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/usuarioDTO"}}}}}},"delete":{"tags":["usuario-controller"],"operationId":"eliminar","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK"}}}},"/api/Partida/{id}":{"get":{"tags":["partida-controller"],"operationId":"getById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/partidaDTO"}}}}}},"put":{"tags":["partida-controller"],"operationId":"actualizar_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Partida"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/partidaDTO"}}}}}}},"/api/Partida/{id}/turno":{"put":{"tags":["partida-controller"],"operationId":"pasarTurno","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK"}}}},"/api/Partida/{id}/ronda":{"put":{"tags":["partida-controller"],"operationId":"iniciarRonda","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK"}}}},"/api/Partida/{id}/guardar":{"put":{"tags":["partida-controller"],"operationId":"guardar","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK"}}}},"/api/Partida/{id}/fase":{"put":{"tags":["partida-controller"],"operationId":"avanzarFase","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK"}}}},"/api/Jugador/{id}":{"get":{"tags":["jugador-controller"],"operationId":"getById_3","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/jugadorDTO"}}}}}},"put":{"tags":["jugador-controller"],"operationId":"update","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"oneOf":[{"$ref":"#/components/schemas/Jugador"},{"$ref":"#/components/schemas/Bot"}]}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/jugadorDTO"}}}}}},"delete":{"tags":["jugador-controller"],"operationId":"delete","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK"}}}},"/api/Usuario/register":{"post":{"tags":["usuario-controller"],"operationId":"crear","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/loginDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/usuarioDTO"}}}}}}},"/api/Usuario/login":{"post":{"tags":["usuario-controller"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/loginDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/usuarioDTO"}}}}}}},"/api/Partida":{"get":{"tags":["partida-controller"],"operationId":"getAll_1","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/partidaDTO"}}}}}}},"post":{"tags":["partida-controller"],"operationId":"crear_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Partida"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"integer","format":"int32"}}}}}}},"/api/Partida/{id}/mover":{"post":{"tags":["partida-controller"],"operationId":"mover","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/moverDTO"}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/api/Partida/{id}/iniciar":{"post":{"tags":["partida-controller"],"operationId":"iniciar","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK"}}}},"/api/Partida/{id}/colocar":{"post":{"tags":["partida-controller"],"operationId":"colocar","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/colocarDTO"}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/api/Partida/{id}/atacar":{"post":{"tags":["partida-controller"],"operationId":"atacar","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/atacarDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultadoAtaque"}}}}}}},"/api/Jugador":{"get":{"tags":["jugador-controller"],"operationId":"getAll_4","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/jugadorDTO"}}}}}}},"post":{"tags":["jugador-controller"],"operationId":"createJugador","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/jugadorDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/jugadorDTO"}}}}}}},"/api/Jugador/partida/{id}":{"get":{"tags":["jugador-controller"],"operationId":"getByPartida","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/jugadorDTO"}}}}}}},"post":{"tags":["jugador-controller"],"operationId":"addJugadorAPartida","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"oneOf":[{"$ref":"#/components/schemas/Jugador"},{"$ref":"#/components/schemas/Bot"}]}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/jugadorDTO"}}}}}}},"/api/Jugador/Partida":{"post":{"tags":["jugador-controller"],"operationId":"crearJugadorEnPartida","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/jugadorPartidaDTO"}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/api/Bot/turno/{idPartida}":{"post":{"tags":["bot-controller"],"operationId":"ejecutarTurnosBots","parameters":[{"name":"idPartida","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}}}}},"/api/Bot/agregar/{idPartida}":{"post":{"tags":["bot-controller"],"operationId":"agregarBots","parameters":[{"name":"idPartida","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/agregarbotDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/ping":{"get":{"tags":["ping-controller"],"summary":"Check healthy of the app","description":"If the app it's alive response pong","operationId":"pong","responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}},"200":{"description":"Successful operation","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/Usuario":{"get":{"tags":["usuario-controller"],"operationId":"listarTodos","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/usuarioDTO"}}}}}}}},"/api/Turno/{idPartida}":{"get":{"tags":["turno-controller"],"operationId":"getTurnoActual","parameters":[{"name":"idPartida","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"integer","format":"int32"}}}}}}},"/api/Turno/{idPartida}/historial":{"get":{"tags":["turno-controller"],"operationId":"getHistorialTurnos","parameters":[{"name":"idPartida","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/turnoDTO"}}}}}}},"/api/Tarjeta":{"get":{"tags":["tarjeta-controller"],"operationId":"getAll","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/tarjetaDTO"}}}}}}}},"/api/Tarjeta/jugador/{id}":{"get":{"tags":["tarjeta-controller"],"operationId":"getJugadorTarjetas","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/tarjetaDTO"}}}}}}}},"/api/Partida/{id}/cargar":{"get":{"tags":["partida-controller"],"operationId":"cargar","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/partidaDTO"}}}}}}},"/api/Partida/reanudar":{"get":{"tags":["partida-controller"],"operationId":"reanudarPartida","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/partidaDTO"}}}}}}},"/api/Partida/partidas/{id}/ganador":{"get":{"tags":["partida-controller"],"operationId":"obtenerGanador","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/jugadorDTO"}}}}}}},"/api/Pais":{"get":{"tags":["pais-controller"],"operationId":"getAll_2","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/paisDTO"}}}}}}}},"/api/Pais/{nombre}/vecinos":{"get":{"tags":["pais-controller"],"operationId":"getVecinosPorNombre","parameters":[{"name":"nombre","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/paisDTO"}}}}}}}},"/api/Pais/{nombre}/tropas":{"get":{"tags":["pais-controller"],"operationId":"getTropasPorNombre","parameters":[{"name":"nombre","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"integer","format":"int32"}}}}}}},"/api/Pais/{id}":{"get":{"tags":["pais-controller"],"operationId":"getById_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/paisDTO"}}}}}}},"/api/Objetivo":{"get":{"tags":["objetivo-controller"],"operationId":"getAll_3","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/objetivoDTO"}}}}}}}},"/api/Objetivo/{id}":{"get":{"tags":["objetivo-controller"],"operationId":"getById_2","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/objetivoDTO"}}}}}}},"/api/Continente":{"get":{"tags":["continente-controller"],"operationId":"getAll_5","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Continente"}}}}}}}}},"components":{"schemas":{"Usuario":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"nombre":{"type":"string"},"contrasenia":{"type":"string"},"nivel":{"type":"integer","format":"int32"},"idTipoJugador":{"type":"string","enum":["BOT","HUMANO"]}}},"usuarioDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"nombre":{"type":"string"},"contrasenia":{"type":"string"},"nivel":{"type":"integer","format":"int32"}}},"Bot":{"type":"object","allOf":[{"$ref":"#/components/schemas/Jugador"},{"type":"object","properties":{"dificultad":{"type":"string","enum":["NOVATO","INTERMEDIO","EXPERTO"]}}}]},"Jugador":{"required":["tipoJugador"],"type":"object","properties":{"id":{"type":"integer","format":"int32"},"nombreJ":{"type":"string"},"color":{"type":"string","enum":["VERDE","ROJO","NEGRO","AZUL","AMARILLO","MAGENTA"]},"estadoJ":{"type":"string","enum":["ACTIVO","ELIMINADO","GANADOR"]},"fichasJ":{"type":"integer","format":"int32"},"objetivo":{"$ref":"#/components/schemas/Objetivo"},"tipoJ":{"type":"string","enum":["BOT","HUMANO"]},"partida":{"$ref":"#/components/schemas/Partida"},"paises":{"type":"array","items":{"$ref":"#/components/schemas/Pais"}},"tarjetas":{"type":"array","items":{"$ref":"#/components/schemas/TarjetaPais"}},"coloresEliminados":{"uniqueItems":true,"type":"array","items":{"type":"string","enum":["VERDE","ROJO","NEGRO","AZUL","AMARILLO","MAGENTA"]}},"numeroCanje":{"type":"integer","format":"int32"},"tipoJugador":{"type":"string"}},"discriminator":{"propertyName":"tipoJugador"}},"Objetivo":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"descripcion":{"type":"string"},"estado":{"type":"boolean"},"tipo":{"type":"string","enum":["CONQUISTA","ELIMINAR"]}}},"Pais":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"nombre":{"type":"string"},"ejercito":{"type":"integer","format":"int32"},"jugador":{"oneOf":[{"$ref":"#/components/schemas/Jugador"},{"$ref":"#/components/schemas/Bot"}]},"partida":{"$ref":"#/components/schemas/Partida"},"vecinos":{"type":"array","items":{"$ref":"#/components/schemas/Pais"}},"vecino":{"$ref":"#/components/schemas/Pais"}}},"Partida":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"estadoPartida":{"type":"string","enum":["EN_JUEGO","GUARDADA","TERMINADA"]},"cantidadJugadores":{"type":"integer","format":"int32"},"ronda":{"type":"integer","format":"int32"},"turno":{"type":"integer","format":"int32"},"fase":{"type":"string","enum":["COLOCACION","ATAQUE","REAGRUPACION"]},"jugadores":{"type":"array","items":{"oneOf":[{"$ref":"#/components/schemas/Jugador"},{"$ref":"#/components/schemas/Bot"}]}},"paises":{"type":"array","items":{"$ref":"#/components/schemas/Pais"}},"objetivos":{"type":"array","items":{"$ref":"#/components/schemas/Objetivo"}},"ganadorTemporal":{"oneOf":[{"$ref":"#/components/schemas/Jugador"},{"$ref":"#/components/schemas/Bot"}]},"jugadorActualId":{"type":"integer","format":"int32"},"jugadorActualNombre":{"type":"string"}}},"TarjetaPais":{"type":"object","properties":{"number":{"type":"integer","format":"int32"},"simbolo":{"type":"string","enum":["INFANTERIA","CABALLERIA","ARTILLERIA"]},"pais":{"$ref":"#/components/schemas/Pais"},"jugador":{"$ref":"#/components/schemas/Jugador"}}},"jugadorDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"nombre":{"type":"string"},"color":{"type":"string","enum":["VERDE","ROJO","NEGRO","AZUL","AMARILLO","MAGENTA"]},"fichas":{"type":"integer","format":"int32"},"tipoJugador":{"type":"string","enum":["BOT","HUMANO"]},"estado":{"type":"string","enum":["ACTIVO","ELIMINADO","GANADOR"]},"paises":{"type":"array","items":{"$ref":"#/components/schemas/paisDTO"}},"objetivo":{"$ref":"#/components/schemas/objetivoDTO"},"tarjetas":{"type":"array","items":{"$ref":"#/components/schemas/tarjetaDTO"}}}},"objetivoDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"descripcion":{"type":"string"},"tipo":{"type":"string","enum":["CONQUISTA","ELIMINAR"]},"estado":{"type":"boolean"}}},"paisDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"nombre":{"type":"string"},"color":{"type":"string","enum":["VERDE","ROJO","NEGRO","AZUL","AMARILLO","MAGENTA"]},"ejercito":{"type":"integer","format":"int32"},"continente":{"type":"string"}}},"partidaDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"estado":{"type":"string","enum":["EN_JUEGO","GUARDADA","TERMINADA"]},"cantidadJugadores":{"type":"integer","format":"int32"},"jugadorActualId":{"type":"integer","format":"int32"},"jugadorActualNombre":{"type":"string"},"ronda":{"type":"integer","format":"int32"},"turno":{"type":"integer","format":"int32"},"fase":{"type":"string","enum":["COLOCACION","ATAQUE","REAGRUPACION"]},"jugadores":{"type":"array","items":{"$ref":"#/components/schemas/jugadorDTO"}}}},"tarjetaDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"simbolo":{"type":"string","enum":["INFANTERIA","CABALLERIA","ARTILLERIA"]},"pais":{"type":"string"}}},"loginDTO":{"type":"object","properties":{"nombre":{"type":"string"},"contrasenia":{"type":"string"}}},"moverDTO":{"type":"object","properties":{"origen":{"type":"string"},"destino":{"type":"string"},"tropas":{"type":"integer","format":"int32"}}},"colocarDTO":{"type":"object","properties":{"tropas":{"type":"integer","format":"int32"},"pais":{"type":"string"}}},"atacarDTO":{"type":"object","properties":{"origen":{"type":"string"},"destino":{"type":"string"}}},"ResultadoAtaque":{"type":"object","properties":{"resultado":{"type":"string"},"dadosAtaque":{"type":"array","items":{"type":"integer","format":"int32"}},"dadosDefensa":{"type":"array","items":{"type":"integer","format":"int32"}}}},"jugadorPartidaDTO":{"type":"object","properties":{"jugadores":{"type":"array","items":{"$ref":"#/components/schemas/jugadorDTO"}},"partidaId":{"type":"integer","format":"int32"}}},"agregarbotDTO":{"type":"object","properties":{"dificultad":{"type":"string","enum":["NOVATO","INTERMEDIO","EXPERTO"]},"cantidad":{"type":"integer","format":"int32"},"colores":{"type":"array","items":{"type":"string","enum":["VERDE","ROJO","NEGRO","AZUL","AMARILLO","MAGENTA"]}}}},"ErrorApi":{"type":"object","properties":{"timestamp":{"type":"string"},"status":{"type":"integer","format":"int32"},"error":{"type":"string"},"message":{"type":"string"}}},"turnoDTO":{"type":"object","properties":{"ronda":{"type":"integer","format":"int32"},"turno":{"type":"integer","format":"int32"},"fase":{"type":"string","enum":["COLOCACION","ATAQUE","REAGRUPACION"]}}},"Continente":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"nombre":{"type":"string"},"cantidad_paises":{"type":"integer","format":"int32"},"conquistado":{"type":"boolean"},"paises":{"type":"array","items":{"$ref":"#/components/schemas/Pais"}}}}}}}